refactor: move comparator for alpha + rc sorting
so it could be usuable for other things too (if needed).

also add an extra test to verify it sorts as expected.
